requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-lru==2.3.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.135.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.17
langchain-openai==1.1.10
langchain-tavily==0.2.17
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.14
multidict==6.7.1
openai==2.26.0
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049137378/work
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.10.2
sniffio==1.3.1
sqlite-vec==0.1.6
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
